1.0.5.0
What’s fixed:
1. Hotkey correct registering and unregistering after saving configuration.

What’s new
1. Single-instance application: app suppresses user to run multiple instances of it
I took the liberty to realize this option because I think it is logical and important for user-friendly behaviour. Application on start checks if it is already running, and if it is true shows user a message about it and exits.
2. No visibility modes
Now app always displays List Window on tray icon click (I think no reaction on tray icon click, like it was in the past in some visibility modes, is a little confusing for user). Also, there are two checkboxes in configuration window: list visibility on hotkey and on input box focus. This is an additional functionality and can always be turned on/off. One note: on List Window Close button click, option “list visibility on input box focus” is turned off like it was in earlier versions.
3. Application always shows List Window on start
Because something’s got to happen when we start a program (and the tray icon can be hidden by OS in unused notification icons area). If “list visibility on input box focus” is turned on, List Window will appear and then app will choose whether it must stay shown or be hidden.